+++ settings.ini
@@ -1,0 +1,28 @@
+# ftpserver configuration file
+#
+# These are all the config parameters with their default values. If not present,
+
+# Max number of control connections to accept
+# max_connections = 0
+MaxConnections = 10
+
+[server]
+# Address and Port to listen on
+ListenAddr="0.0.0.0:2121"
+
+# Public host to expose in the passive connection
+# public_host = ""
+
+# Idle timeout time
+IdleTimeout = 900
+
+# Data port range from 10000 to 15000
+[server.dataPortRange]
+Start = 2122
+End = 2200
+
+[[users]]
+User="admin"
+Pass="admin"
+Dir="/"
+
